♻️[Refactor] BaseEntity 패키지를 common.entity로 이동
