+++ src/main/java/com/seohaeng/backend/domain/place/entity/Review.java
@@ -1,6 +1,6 @@
 package com.seohaeng.backend.domain.place.entity;
 
-import com.seohaeng.backend.domain.common.BaseEntity;
+import com.seohaeng.backend.domain.common.entity.BaseEntity;
 import com.seohaeng.backend.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
