+++ src/main/java/com/seohaeng/backend/domain/readingSpot/entity/ReadingSpot.java
@@ -1,6 +1,6 @@
 package com.seohaeng.backend.domain.readingSpot.entity;
 
-import com.seohaeng.backend.domain.common.BaseEntity;
+import com.seohaeng.backend.domain.common.entity.BaseEntity;
 import com.seohaeng.backend.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
